requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.1
WTForms==3.1.2
email-validator==2.1.0.post1
python-dateutil==2.8.2
Markdown==3.5.2
boto3==1.34.69
aws-wsgi==0.2.7 